flawed fluent API
