+++ src/NServiceBus.WindowsPerformanceCounters/NServiceBus.WindowsPerformanceCounters.csproj
@@ -57,6 +57,7 @@
     <Compile Include="Counters\PerformanceMonitorUsersInstaller.cs" />
     <Compile Include="CriticalTime\CriticalTimeMonitoring.cs" />
     <Compile Include="Metrics.cs" />
+    <Compile Include="MetricsExtensionsForPerformanceCounters.cs" />
     <Compile Include="PerformanceCounters.cs" />
     <Compile Include="PerformanceCountersExtensions.cs" />
     <Compile Include="PerformanceSettings.cs" />
